Fix flakey specs in `ChangeLeadProvider`
They are failing as we are in December and we create future-dated
training periods, however there are no schedules created for the
next year.

Run all tests mid-year to avoid rolling over into the next year.

I tried creating schedules for the next contract period instead, but
that cause further failures.