Sticking plaster fix for flakey spec
This spec fails because the expression of interest is on a different
contract period year to the schedule found in the
`TrainingPeriods::Create` service.

I haven't got to the bottom of what's going wrong here, but I think
its an issue in one of the services rather than just the test setup.