Split `status` into `payment_status` and `clawback_status`
Having a single `status` on a declaration linked to both a
`payment` and `clawback` statement can be confusing when the
declaration is in a `clawback` state when evaluating it from the
perspective of the `payment` statement (as the declaration will be
`awaiting_clawback` or `clawed_back` now that it has moved on from
the `paid` status).

To better reflect the status relative to the statements we are
splitting the `status` into `payment_status` (relevant to the
`payment_statement`) and `clawback_status` (relevant to the
`clawback_statement`).

Given how we look at/query declarations and statement items in
ECF this should support everything cleanly. The only
complexity/muddiness comes in when we serialize the overall state of
a declaration for the API response. We will then need to return the
`latest_status`, which will be the `clawback_status` if set and
falling back to the `payment_status`.