Split API seeds from development seeds
The API seeds contain a lot of extra seed data that isn't relevant
to other teams working on RECT.

Remove the API-specific seed data from the review apps.

Change the sandbox seeds to be generic API seeds (this will have the
benefit of the review apps mirroring sandbox to a degree).

Drop redundant `teachers` seed from API seeds (as we create teachers
with histories; these teachers never get any training periods so are
a bit pointless).

Ensure all API seeds will only run once (if ran a second time they
will be skipped).

Add schdules and milestones to API seeds as these were missing.

Run the API seeds in CI when deploying a PR with the `api-seeds`
label or when an API team member raised the PR.