Add API seed data for school transfers
This builds on #1860 and #1816 to add API seed data for school transfers.

Now, there will be relevant data in sandbox and API review apps to facilitate
testing the `/transfers` endpoints.

The seed file is based mostly on the original work in #1816, except now we
re-use existing teachers if they exist.