Add an end to end test for one induction period
This test covers the pulling of ECF1 data using ECF1TeacherHistory,
the conversion to ECF2TeacherHistory via TeacherHistoryConverter, and
the saving in ECF2's database.

It doesn't test every last bit of info, that's already covered in depth
by the tests for the 3 objects named above, this is making sure they're
working nicely together.

Co-authored-by: tonyheadford <tony@binarycircus.com>