Close/open training periods on mentor completion
When a we get a completed, billable declaration submitted for a mentor we
wantto also close off their latest/ongoing training period for that lead
provider.

Where possible we will use the `declaration_date` as the `finished_on`.
If,however, that date is before the training period `started_on` we will close
it the day after it started.

When a billable, completed declaration is voided we 'uncomplete' the mentor
andwe also want to resume  their training period with the lead provider (unless
ithas been resumed through another avenue). To do this, we will create a new
training period that starts the day their previous period finished and it will
mirror the `finished_on` of the school period.