Add `api_updated_at` to `Declaration`
We need to be able to maintain a separate `updated_at` field for the API response in order to support filtering by `updated_since`.

Add `api_updated_at` to `Declaration`.

Add declarative updates to touch the `api_updated_at` when a serialized field
changes. Note that when a school partnership changes delivery partner we could
end up with a lot of declarations touched, but as this isn't expensive like the
metadata calculations I think we'll be ok. We could look at optimising the
touches to use an `update_all` when its used against a collection if we find
its a bottleneck.