Introduce `API::Declarations::Void` service
In ECF1, the service either voided or clawed back the declaration depending on
whether the declaration was `paid` or not.

We've decided to simplify this by splitting the service into two distinct
services for voiding and clawing back declarations respectively.

This introduces the `Void` service.

Similarly to the `API::Teachers` services, we have split the validation logic
and actions into separate services.

An optional `voided_by_user_id` can be provided so we can track when
declarations are voided by Admin users etc.