Implement void declarations endpoint
Before, the participant declarations `/void` endpoint always returned
`:not_allowed`.

Now that the services exist, along with the declarations query and serializer,
we can build out the endpoint.

This implements the endpoint, with some simple logic to determine whether we
should clawback or void a declaration.

As part of this change, the route has been redefined as a member of
`/declarations` so we can use the same `api_id` param.

The shared "API update" example has also been tweaked slightly so the test for
malformed request bodies can be skipped when an endpoint does not accept a
request body.