Optimise `Declarations::Query` lead provider filtering
The original query was quite easy to follow, but turned out to be very
inefficient; the `previous_declarations_for_lead_provider` generates a subquery
for every row in the scope and contains multiple joins and `exists` checks,
which is very expensive. It also results in multiple full table scans and isn't
able to utilise indexes efficiently.

The updated query, while much more complex and difficult to follow, provides a
more efficient query plan and avoids unnecessery joins.

Running the query on sandbox seed data (~5k declarations) was taking ~3s with
the original query, with the updated one its around 150ms.