Prevent `nil` dynamic request content raising errors
When voiding declarations for Capita we can't find a declaration that we can
void because there are no output fee statements for the lead provider.

In an ideal world we would be able to exclude lead providers from certain
endpoint checks, but as time is of the essense and this code is short lived
I'm opting for a quick fix.