Clear API seeds when deploying to review
I haven't been able to figure out how, but somehow there is the non-api seed
data in the review app after deploying via CI. As we check for existing data in
the API seeds, it causes the majority of them not to run.

Truncate the database before running the API seeds in CI to ensure they are
always ran. This also mirror s the dev seeds (which `replant` on each deploy).