Optimise `Metadata::Handlers::Teacher` school transfers check
Before, we built the full school transfer history to determine whether or not a
given teacher had school transfers in the metadata handler.

We don't need the full history, though, since we only care if there is at least
one transfer.

This refactors the history class so we can return early if a school transfer
exists.