Anonymise test cases generated from real data
There are certain records in ECF1 where the history is complicated. We
want to be sure our migrators handle it properly, so importing the exact
setup and verifying we build the ECF2 representation of the data is a
sensible check.

There is some identifying information though. In addition to relying on
developers to strip it out, this functionality changes the identifying
bits; namely:

* the name
* the TRN
* the school

That leaves delivery partner and lead provider, along with the ECF1
IDs. These don't provide enough information to work out an identity.

We should confirm the above suspicion with the team and can replace the
lead providers and delivery partners in the same way we do schools if
needed, but that might make the tests not resemble reality enough.