Add delivery partner and preferred identity to anonymisation
As we'll be using this generator with real data we want to ensure
nothing's left that will help identify people.

Clearly the email address is personal info, so we'll strip that, but
delivery partner also (potentially) poses a risk - albeit minimal.

Smaller delivery partners might only work with a small number of
teachers, so in conjunction with other information it could (in theory)
be used to triangulate a person. The other information required is held
by DfE, so it's incredibly unlikely a person without it could perform
the task.

Lead providers work with larger numbers of teachers and being able to
see them will help us work with the data, so they'll be left.